+++ src/hnsw/usearch_storage.hpp
@@ -8,7 +8,7 @@ extern "C" {
 
 uint32_t UsearchNodeBytes(const metadata_t *metadata, int vector_bytes, int level);
 void     usearch_init_node(
-        metadata_t *meta, char *tape, unsigned long key, uint32_t level, uint32_t slot_id, void *vector, size_t vector_len);
+        metadata_t *meta, char *tape, unsigned long key, uint32_t level, uint64_t slot_id, void *vector, size_t vector_len);
 
 char *extract_node(char             *data,
                    uint64_t          progress,
